Adding structure for Operations Manual
for:
bp openstack-operations-manual

Adding a structure and a bit of padding text for an eventual
OpenStack operations manual, based on:
http://etherpad.openstack.org/EssexOperationsGuide

patch2: adding a random README to make the figures directory stick

patch3: changed license in book from apache2 to cc-by-nc-sa
but also asking Pearson if that is their intent.

patch4: Pearson okay with cc-by-sa for these chapters.

Change-Id: I34a2e36af98b59f78c828bb3b265ca736ad2df0b