Quick pass fixes for ops guide old folsom links
There were a number of links in the ops guide that went to old
folsom pages for doc references.

Many of these can now point to the configuration reference, or
havana docs safely.

Few minor fixes to the text and adding status labels after review.

As to when to merge this - your call reviewerinos :)

Change-Id: I8844dcd9e7eb21584167f59fec2ed44bd1164761