Replace keystone service and endpoint with catalog
This commit replaces the use of keystone service and keystone endpoint
with the "keystone catalog" command. The main reason for doing this is
because the service/endpoint commands only work when the Keystone
catalog has a SQL backend. It does not work with the templated catalog
backend. The templated catalog backend is becoming more popular in
environments due to the simplicity of a single text file as well as
the decreased load seen on databases by not querying for endpoints.

Change-Id: I17f9dab499e056cb198f79325c0213399f729a51