Remove all mostly untranslated PO files
We only import PO files that are at least 75 % translated (8 % in
common), so we can delete now all other PO files. The patch only
removes the mostly untranslated files. Once a file becomes mostly
translated, the bot will import it again.

Japanese is only translated at 47 % but I've left it in, since we're
currently building the guide in Japanese.

Partial-Bug: #1317794

Change-Id: I90416726e72e1daeda4e5aba41972a74c2cb8647