Update wording to make more sense for metadata changes.
The wording is updated to say 'flavor' instead of 'instance'
in a lot of places where 'flavor type' makes more sense to
see when reading through the page. The wording 'instance type'
is not always clear, and since the changes are for flavor
metadata, this made more sense.

(Fixed a couple 'extra specs' to be 'extra_specs' like the
rest of the places it is.)

Change-Id: I702481e6b67536852efc3a0747b29be811cf792f