Fixes for post-Juno world
Now that Juno has been released, there are some sections
in this guide that need to be updated.

This patch has a first hash of the "working with roadmaps"
appendix. It mainly focuses on getting things up to date
to around Juno Release. A later patch will add more forward-looking
statements toward Kilo.

In addition, a couple of small updates to the usecases chapter.

Change-Id: I2529072c0849467d98e39180cfdb07081e00dd02