Removed line suggesting a future migration path from nova-network.
The line that was removed stated, “Current thinking is that upgrades without instance downtime will be available in the Kilo release.” It is still unknown whether or not there will be a migration path from nova-network. We are removing this sentence which hints to the possibility of a migration path in the future, specifically Kilo.

Change-Id: I9428ce50942d7c1b7e4efaa5129d5e80edb2d74d
Closes-Bug: #1478136