Add Image Signing Instructions
Add instructions in the Operations Guide about how to use the new
image signing feature, which was released in Compute during Mitaka.

Note that Devstack-based steps to create asymmetric keypair and cert
will be added to the user-guide here
http://docs.openstack.org/developer/barbican/api/,
but enterprise-capable steps will probably need to go to TBD
Castellan page.

Closes-Bug: #1507623
Change-Id: I208e65e47617f3054b97ebf468f2dff98feeedeb